# docs/swagger.yaml
basePath: /
definitions:
...
  handler.HTTPError:
    properties:
      error:
        type: string
    type: object
  handler.HTTPMessage:
    properties:
      message:
        type: string
    type: object
...
  handler.PropertyShortView:
    properties:
      price:
        type: number
      thumbnail:
        type: string
      title:
        type: string
      value:
        type: string
    type: object
...
  handler.VariationShortView:
    properties:
      thumbnail:
        type: string
      title:
        type: string
    type: object
...
  handler.WishView:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      price:
        type: number
      properties:
        items:
          $ref: '#/definitions/handler.PropertyShortView'
        type: array
      thumbnail:
        type: string
      title:
        type: string
      uuid:
        type: string
      variation:
        $ref: '#/definitions/handler.VariationShortView'
        type: object
    type: object
  handler.WishWrapperView:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      price:
        type: number
      thumbnail:
        type: string
      title:
        type: string
      uuid:
        type: string
    type: object
  handler.WishesView:
    items:
      $ref: '#/definitions/handler.WishView'
    type: array
  handler.WithRequest:
    properties:
      categoryId:
        type: integer
      uuid:
        type: string
    type: object
...
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: GoShop full featured api, see documentation https://docs.google.com/document/d/1VlkAYTqZG9oGvZxnSNDpcuafU-msOb2Qm2FevMJkA7w/edit
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: "1.0"
paths:
...
  /api/v1/account/wishes:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.WishesView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Get account wishes
      tags:
      - account
      - frontend
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: option
        required: true
        schema:
          $ref: '#/definitions/handler.WithRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.WishWrapperView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Create account wish
      tags:
      - account
      - frontend
  /api/v1/account/wishes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Wish ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Delete account wish
      tags:
      - account
      - frontend
    get:
      consumes:
      - application/json
      parameters:
      - description: Wish ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.WishWrapperView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Get account wish
      tags:
      - account
      - frontend
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"